refactor: components 디렉토리 구조 변경
- 공통 컴포넌트의 경우 하위 디렉토리인 common으로 이동
