+++ client/src/__tests__/components/inputs/SelectPaymentMethod.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 import dummyOptions from '../../../__dummy-data__/components/inputs/dummyOptions';
-import SelectPaymentMethod from '../../../components/inputs/select-payment-method/SelectPaymentMethod';
+import SelectPaymentMethod from '../../../components/common/inputs/select-payment-method/SelectPaymentMethod';
 
 describe('SelectPaymentMethod 테스트', () => {
   const spy = jest.fn();
