+++ client/src/components/common/color-picker/ColorPicker.stories.tsx
