+++ client/src/components/common/color-picker/ColorPicker.tsx
