+++ client/src/components/common/profile-image/ProfileImage.stories.tsx
