+++ client/src/components/common/profile-image/ProfileImage.tsx
