+++ client/src/components/common/small-accountbook-item/SmallAccountbookItem.stories.tsx
