+++ client/src/components/common/transactions/all-transaction-container/AllTransactionContainer.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import AllTransactionCotainer from './AllTransactionContainer';
-import { transactions } from '../../../__dummy-data__/components/transactions/dummyData';
+import { transactions } from '../../../../__dummy-data__/components/transactions/dummyData';
 import styled from 'styled-components';
 
 export default {
