+++ client/src/components/common/transactions/day-transaction-container/DayTransactionContainer.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import DayTransactionCotainer from './DayTransactionContainer';
-import { transactions } from '../../../__dummy-data__/components/transactions/dummyData';
+import { transactions } from '../../../../__dummy-data__/components/transactions/dummyData';
 
 export default {
   title: 'transactions/day-transaction-container/DayTransactionContainer',
